Move accessibility.h / accessibility.cpp into their own folder. JTRUE/JFALSE instead of true/false. Tweak wording of one caption
