+++ TheForceEngine/Captions/subtitles.txt
@@ -37,8 +37,8 @@ ex-med1.voc "[Big Explosion]"
 ex-tiny.voc "[Bang]"
 swim-in.voc "[Splash]"
 gamor-3.voc "[Porcine grunting]"
-reeyee1.voc "[Yelling in alien tongue]"
-reeyee-1.voc "[Yelling in alien tongue]"
+reeyee1.voc "[Alien yelling]"
+reeyee-1.voc "[Alien yelling]"
 probe-1.voc "[Staccato droid speech]"
 //probalm.voc "[Circuits overloading]"
 bossk-1.voc "[Alien hissing]"
